+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v1.18.0
+
 Added:
 * Add `chronosphere_service_attributes` resource for setting up service attributes.
 * Add `skip_on_conflict` field to `chronosphere_rollup_rule` resource.